14pr refactor auth, rules

+++ controllers/cards.js
@@ -3,25 +3,16 @@ const Card = require('../models/card'); // данные
 const {
   OK_STATUS_CODE,
   HTTP_CREATED_STATUS_CODE,
-  // HTTP_BAD_REQUEST_STATUS_CODE,
-  // HTTP_UNAUTHORIZED_STATUS_CODE,
-  // NOT_FOUND_PAGE_STATUS_CODE,
-  // SERVER_ERROR_STATUS_CODE,
 } = require('../utils/errors');
 const NotFoundError = require('../errors/not-found-err');
 const BadRequestError = require('../errors/bad-request-err');
-// const UnauthorizedError = require('../errors/unauthorized-err');
+const UnauthorizedError = require('../errors/unauthorized-err');
 
 // Получение cards
 const getCards = (req, res, next) => {
   Card.find({})
     .then((cards) => res.status(OK_STATUS_CODE).send(cards))
     .catch(next);
-  //   () => {
-  //   res
-  //     .status(SERVER_ERROR_STATUS_CODE)
-  //     .send({ message: 'Произошла ошибка на сервере при запросе карточек' });
-  // });
 };
 
 // создание card
@@ -34,54 +25,40 @@ const createCard = (req, res, next) => {
       if (err.name === 'ValidationError') {
         throw new BadRequestError('При создании карточки переданы некорректные данные');
         // res.status(HTTP_BAD_REQUEST_STATUS_CODE).send({
-        //   message: 'При создании карточки переданы некорректные данные',
-        // });
       }
       next(err);
-      // else {
-      //   res
-      //     .status(SERVER_ERROR_STATUS_CODE)
-      //     .send({ message: 'Произошла ошибка создания карты на сервере' });
-      // }
     });
 };
 
 // удаление card
 const deleteCard = (req, res, next) => {
   const { cardId } = req.params;
-  // console.log(req.params);
-  // const owner = req.user._id; // используем req.user
-  Card.findByIdAndRemove(cardId)
+  console.log(req.params);
+  const someOwner = req.user._id; // используем req.user
+  console.log(someOwner);
+  Card
+    .findById(cardId)
     .then((card) => {
       if (!card) {
         throw new NotFoundError('Не найдена карточка с таким id');
-        // res
-        //   .status(NOT_FOUND_PAGE_STATUS_CODE)
-        //   .send({ message: 'Не найдена карточка с таким id' });
-        // return;
       }
-      // if (cardOwner !== owner) {
-      //   res
-      //     .status(HTTP_UNAUTHORIZED_STATUS_CODE)
-      //     .send({ message: 'Карточку может удалить только ее автор' });
-      //   return;
-      // }
-      res.status(OK_STATUS_CODE).send(card);
+      const { owner: cardOwner } = card;
+
+      if (cardOwner.valueOf() !== someOwner) {
+        throw new UnauthorizedError('Карточку может удалить только ее автор');
+      }
+      Card.findByIdAndRemove(cardId)
+        .then((card) => {
+          res.status(OK_STATUS_CODE).send(card);
+        })
+        .catch((err) => {
+          if (err.name === 'CastError') {
+            throw new BadRequestError('При попытке удаления карточки переданы некорректные данные');
+          }
+          next(err);
+        });
     })
-    .catch((err) => {
-      if (err.name === 'CastError') {
-        throw new BadRequestError('При попытке удаления карточки переданы некорректные данные');
-        // res.status(HTTP_BAD_REQUEST_STATUS_CODE).send({
-        //   message: 'При попытке удаления карточки переданы некорректные данные',
-        // });
-      }
-      next(err);
-      // else {
-      //   res.status(SERVER_ERROR_STATUS_CODE).send({
-      //     message: 'Произошла ошибка на сервере при удалении карточки',
-      //   });
-      // }
-    });
+    .catch(next);
 };
 
 // Поставить лайк на card
@@ -92,26 +69,14 @@ const addLikeCard = (req, res, next) => {
     .then((card) => {
       if (!card) {
         throw new NotFoundError('Не найдена карточка с таким id');
-        // res
-        //   .status(NOT_FOUND_PAGE_STATUS_CODE)
-        //   .send({ message: 'Не найдена карточка с таким id' });
-        // return;
       }
       res.status(OK_STATUS_CODE).send(card);
     })
     .catch((err) => {
       if (err.name === 'CastError') {
         throw new BadRequestError('При попытке поставить лайк карточке переданы некорректные данные');
-        // return res.status(HTTP_BAD_REQUEST_STATUS_CODE).send({
-        //   message:
-        //     'При попытке поставить лайк карточке переданы некорректные данные',
-        // });
       }
       next(err);
-      // return res.status(SERVER_ERROR_STATUS_CODE).send({
-      //   message:
-      //     'Произошла ошибка на сервере при постановке лайка этой карточки',
-      // });
     });
 };
 
@@ -123,25 +88,14 @@ const deleteLikeCard = (req, res, next) => {
     .then((card) => {
       if (!card) {
         throw new NotFoundError('Не найдена карточка с таким id');
-        // res
-        //   .status(NOT_FOUND_PAGE_STATUS_CODE)
-        //   .send({ message: 'Не найдена карточка с таким id' });
-        // return;
       }
       res.status(OK_STATUS_CODE).send(card);
     })
     .catch((err) => {
       if (err.name === 'CastError') {
         throw new BadRequestError('При попытке убрать лайк карточки переданы некорректные данные');
-        // return res.status(HTTP_BAD_REQUEST_STATUS_CODE).send({
-        //   message:
-        //     'При попытке убрать лайк карточки переданы некорректные данные',
-        // });
       }
       next(err);
-      // return res
-      //   .status(SERVER_ERROR_STATUS_CODE)
-      //   .send({ message: 'Произошла ошибка при удалении лайка этой карточки' });
     });
 };
